Renaming the validator_cache module to nym_contract_cache
